Created components directory and moved app to there. Also, updated index.js.
